refactor: implement Port interface directly on adapter
